Add in Histogram1D from reaction_plane_fit, refactor in histogram module
Moved the `Histogram1D` object from the reaction_plane_fit package into
pachyderm so that it can be used in other projects. Since this added a
good deal of histogram functionality (some of which more or less
duplicated functionality in the `utils` module), I refactored the histogram
functionality into a new `histogram` module.

This is a larger commit than I'd like, but it ended up being rather
convenient, so I'd just going with it.