Fix overriding due to ruamel.yaml change
They introduced a nice change to preserve anchors for basic types (int,
float, str, bool), but this also means that these types (sometimes) have
anchors. Thus, our override intuition went wrong.

Note that it is unclear how to update the values while preserving the
anchors, so for now we apply a hack (note that this is no worse than our
previous capabilities). Note also that while floats are always reconstructed
as `ScalarFloat`, this isn't always the case for other types (int, etc).
This is why tests failed for float, but worked for int. So some issues
there need to be worked out.

In summary, the hack allows normal operations (and can potentially be
improved in the future). This commit also adds better protection for
these type of issues, so they should be more apparent when they go wrong.