Appears to fix the test issues locally
It appears to be some unexplained ROOT memory issues that is caused by
raising an exception. It appears not to occur anymore if the ROOT file
is closed when the exception is raised (which is good practices anyway).
Since the tests appear to work locally, it's not worth investigating
further (assuming that this works on Travis...).