Simplify projecting a single observable
Projecting to dictionaries was nice because multiple objects could refer
to the same dictionary, but such an interface requires a number of
temporary dictionaries when used with single objects.  Now, one can
assign the output to an attribute of a given object, thereby simplifying
the user experience for projecting a single object.

Note that the interface supports both the single object and the
dictionary outputs