Add find_bin method for getting bin from value
It's unclear how much use this will get, but if we provide more
functionality in the future, it could be quite useful. So we just do it
now.

The corresponding tests also provide a new and improved fixture with a
`Histogram1D` that is simple, but has interesting features in terms of
non-uniform binning and weighted filling