Improve numpy encoding when written to YAML
By fully encoding the values in bytes, it should avoid the issue where
the least significant digits are fluctuating from one calculation to the
next, which makes it appears as if there are changes to the files when
nothing has actually changed.

The encoded values are not transparent when viewed in the file, but if
it fixes the encoding issues, it will be worth it.