Fix binned log likelihood to agree with ROOT
The actual value of the log likelihood has a major impact on the
covariance matrix (and/or equivalently(?), the parameter errors
and the estimated distance from the minimum), which then translates
to differences in the point-by-point errors. By modifying the overall
normalization of the log likelihood function, it now agrees with the
ROOT calculation. Further, we now explicitly support whether to utilize
the histogram weights when performing the log likelihood fit.