Add class for parallel downloading a list of file pairs
This should be used for simpler cases where we can easily enumerate the
files by hand. There is no entry point because the user is expected to
write a simple script to call the functions on their own.

We also correspondingly make the `download` function that actually directs
filling the queue and downloading public.